Change types to explicit avr format

+++ arduino/chomp/telem.cpp
@@ -9,28 +9,29 @@ bool cts(){
   return (digitalRead(XBEE_CTS) == LOW);
 }
 
-void write_int(char* packet, int n){
-  for (int i = 0; i < 4; i++){
+// I think int is same as short in avrland
+void write_int(char* packet, int16_t n){
+  for (uint8_t i = 0; i < 4; i++){
     packet[i] = 0xFF & n >> (i*8); 
   }
 }
 
-void write_ulong(char* packet, unsigned long n){
-  for (int i = 0; i < 4; i++){
+void write_ulong(char* packet, uint32_t n){
+  for (uint8_t i = 0; i < 4; i++){
     packet[i] = 0xFF & n >> (i*8); 
   }
 }
 
-void write_short(char* packet, short n){
-  for (int i = 0; i < 2; i++){
+void write_short(char* packet, int16_t n){
+  for (uint8_t i = 0; i < 2; i++){
     packet[i] = 0xFF & n >> (i*8); 
   }
 }
 
 void write_float(char* packet, float n){
   // get access to the float as a byte-array:
-  byte * data = (byte *) &n;
-   for (int i = 0; i < 4; i++){
+  uint8_t * data = (uint8_t *) &n;
+   for (uint8_t i = 0; i < 4; i++){
     packet[i] = data[i];
   }
 }
@@ -40,8 +41,8 @@ void write_terminator(char* packet){
   packet[1] = 0x55;
 }
 
-void send_sensor_telem(unsigned long loop_speed, float pressure){
-  const unsigned int LEN = 11;
+void send_sensor_telem(uint32_t loop_speed, float pressure){
+  const uint8_t LEN = 11;
   char packet[LEN];
   write_terminator(packet);
   packet[3] = 1; // packet id
@@ -59,7 +60,7 @@ void sendLeddarTelem(Detection* detections, unsigned int count, LeddarState stat
 //  char packet_id = 2;
 //  Xbee.write(packet_id);
 //  write_int(state);
-//  int j=0;
+//  uint8_t j=0;
 //  for (int i = 0; i < 16; i++){
 ////    Xbee.print(detections[i].Segment);
 ////    Xbee.print("/");
